Split EventTime class into fluent/event_time.rb
